+++ src/main/java/com/example/dreamday/domain/Categoria.java
@@ -2,21 +2,21 @@ package com.example.dreamday.domain;
 
 public class Categoria {
 
-    private Integer id;
+    private Long id;
     private String nome;
 
     public Categoria() {}
 
-    public Categoria(Integer id, String nome) {
+    public Categoria(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
